% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{computemy}
\alias{computemy}
\title{Compute a process and output results to a table within a database schema.}
\usage{
computemy(con, object, schema, table)
}
\arguments{
\item{con}{Connection string to the database.}

\item{object}{Pipeline object to be computed.}

\item{schema}{Schema within the database. In DCVTS, each round is a different schema.}

\item{table}{Connection to the specific table within the database schema.}
}
\description{
Compute a process and output results to a table within a database schema.
}
\details{
\code{computemy()} also checks if a table previously exists and if it does, it removes the table and writes the new computation to the table. It also prints how long it to perform the computation within the database.
}
\examples{
\dontrun{
computemy(con, "r2", "responses")
computemy(con, "r2", "summary")
}
}
