% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{dbWriteTablemy}
\alias{dbWriteTablemy}
\title{Write the object to a table within a database schema.}
\usage{
dbWriteTablemy(con, object, schema, table)
}
\arguments{
\item{con}{Connection string to the database.}

\item{object}{R object to be written to the database.}

\item{schema}{Schema within the database. In DCVTS, each round is a different schema.}

\item{table}{Connection to the specific table within the database schema.}
}
\description{
Write the object to a table within a database schema.
}
\details{
Any previous table is overwritten.
}
\examples{
\dontrun{
dbWriteTablemy(con, "r2", "responses")
dbWriteTablemy(con, "r2", "summary")
}
}
