% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{tblmy}
\alias{tblmy}
\title{Nicer connection syntax to a table within a database schema.}
\usage{
tblmy(con, schema, table)
}
\arguments{
\item{con}{Connection string to the database.}

\item{schema}{Schema within the database. In DCVTS, each round is a different schema.}

\item{table}{Connection to the specific table within the database schema.}
}
\value{
Connection to the corresponding table is returned.
}
\description{
Nicer connection syntax to a table within a database schema.
}
\examples{
\dontrun{
tblmy(con, "r2", "responses")
tblmy(con, "r2", "summary")
}
}
